moved plugin-sbt-testrunner files
